Improve handling of ASN1 types
Asn1Type now implements hashCode() and equals()
  Two Asn1Type's are considered equal if they represent the same native
enum. As there is only one instance per enum value, this should be
equivalent to testing object identity, but this is just in case.

Asn1Type now has a name field, which is the same string that is used to
query for that type in Asn1Types.getAsn1TypeByName

Asn1Types can now be used to iterate over all supported types